requirements.txt
langchain
langchain-experimental
langchain-core>=0.3.15
langchain-openai
langchain-groq
langchain-tavily
langchainhub
langgraph
streamlit
python-dotenv
pypdf
bs4
arxiv
pymupdf
matplotlib
langchain-text-splitters
langchain-chroma
langchain-community
ipykernel
langgraph-cli[inmem]